Calculate confidence intervals of proportion or difference between proportions

Calculate confidence intervals of proportion or difference between proportions

Usage

propCI(
  x,
  y,
  n,
  p,
  n1,
  n2,
  p1,
  p2,
  P = 0,
  alpha = 0.05,
  digits = 2,
  alternative = "two.sided"
)

Arguments

x

A vector

y

A vector

n, n1, n2

integer sample size

p, p1, p2, P

Numeric proportion

alpha

numeric confidence level

digits

integer indicating the number of decimal places

alternative

A character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less".

Examples

propCI(acs$sex)
#> $data
#> # A tibble: 857 × 1
#>    value 
#>    <chr> 
#>  1 Male  
#>  2 Female
#>  3 Female
#>  4 Female
#>  5 Male  
#>  6 Female
#>  7 Male  
#>  8 Male  
#>  9 Female
#> 10 Male  
#> # … with 847 more rows
#> 
#> $result
#>   alpha   n  df         p P         se critical         ME     lower     upper
#> 1  0.05 857 856 0.6651109 0 0.01612156 1.959964 0.03159768 0.6335132 0.6967085
#>                       CI        z pvalue alternative
#> 1 0.67 [95CI 0.63; 0.70] 41.25598      0   two.sided
#> 
#> $call
#> propCI(x = acs$sex)
#> 
#> attr(,"measure")
#> [1] "prop"
propCI(acs$sex,acs$DM)
#> $data
#> # A tibble: 857 × 2
#>    x      y    
#>    <chr>  <chr>
#>  1 Male   Yes  
#>  2 Female No   
#>  3 Female No   
#>  4 Female No   
#>  5 Male   Yes  
#>  6 Female Yes  
#>  7 Male   Yes  
#>  8 Male   Yes  
#>  9 Female Yes  
#> 10 Male   Yes  
#> # … with 847 more rows
#> 
#> $result
#>   alpha        p1        p2  n1  n2   DF         pd         se critical
#> 1  0.05 0.3972125 0.3333333 857 857 1712 0.06387921 0.02320969 1.959964
#>           ME      lower     upper                     CI   ppooled   sepooled
#> 1 0.04549015 0.01838906 0.1093694 0.06 [95CI 0.02; 0.11] 0.3652729 0.02326092
#>          z      pvalue alternative
#> 1 2.746203 0.006028938   two.sided
#> 
#> $call
#> propCI(x = acs$sex, y = acs$DM)
#> 
#> attr(,"measure")
#> [1] "propdiff"
propCI(n=1600,p=0.4,alpha=0.01)
#> $data
#> # A tibble: 1 × 1
#>   value
#>   <lgl>
#> 1 NA   
#> 
#> $result
#>   alpha    n   df   p P         se critical         ME     lower     upper
#> 1  0.01 1600 1599 0.4 0 0.01224745 2.575829 0.03154734 0.3684527 0.4315473
#>                       CI        z        pvalue alternative
#> 1 0.40 [99CI 0.37; 0.43] 32.65986 5.804914e-234   two.sided
#> 
#> $call
#> propCI(n = 1600, p = 0.4, alpha = 0.01)
#> 
#> attr(,"measure")
#> [1] "prop"
propCI(n=100,p=0.73,P=0.8,alpha=0.01)
#> $data
#> # A tibble: 1 × 1
#>   value
#>   <lgl>
#> 1 NA   
#> 
#> $result
#>   alpha   n df    p   P   se critical        ME     lower     upper
#> 1  0.01 100 99 0.73 0.8 0.04 2.575829 0.1030332 0.6269668 0.8330332
#>                       CI     z     pvalue alternative
#> 1 0.73 [99CI 0.63; 0.83] -1.75 0.08011831   two.sided
#> 
#> $call
#> propCI(n = 100, p = 0.73, P = 0.8, alpha = 0.01)
#> 
#> attr(,"measure")
#> [1] "prop"
propCI(n1=400,n2=300,p1=0.4,p2=0.3,alpha=0.1)
#> $data
#> # A tibble: 1 × 2
#>   x     y    
#>   <lgl> <lgl>
#> 1 NA    NA   
#> 
#> $result
#>   alpha  p1  p2  n1  n2  DF  pd         se critical         ME      lower
#> 1   0.1 0.4 0.3 400 300 698 0.1 0.03605551 1.644854 0.05930604 0.04069396
#>      upper                     CI   ppooled   sepooled       z      pvalue
#> 1 0.159306 0.10 [90CI 0.04; 0.16] 0.3571429 0.03659625 2.73252 0.006285182
#>   alternative
#> 1   two.sided
#> 
#> $call
#> propCI(n1 = 400, n2 = 300, p1 = 0.4, p2 = 0.3, alpha = 0.1)
#> 
#> attr(,"measure")
#> [1] "propdiff"
propCI(n1=100,n2=200,p1=0.38,p2=0.51,alpha=0.01)
#> $data
#> # A tibble: 1 × 2
#>   x     y    
#>   <lgl> <lgl>
#> 1 NA    NA   
#> 
#> $result
#>   alpha   p1   p2  n1  n2  DF    pd         se critical        ME      lower
#> 1  0.01 0.38 0.51 100 200 298 -0.13 0.06004582 2.575829 0.1546678 -0.2846678
#>        upper                       CI   ppooled   sepooled         z     pvalue
#> 1 0.02466777 -0.13 [99CI -0.28; 0.02] 0.4666667 0.06110101 -2.127624 0.03336824
#>   alternative
#> 1   two.sided
#> 
#> $call
#> propCI(n1 = 100, n2 = 200, p1 = 0.38, p2 = 0.51, alpha = 0.01)
#> 
#> attr(,"measure")
#> [1] "propdiff"
propCI(n1=100,n2=200,p1=0.38,p2=0.51,alpha=0.01,alternative="less")
#> $data
#> # A tibble: 1 × 2
#>   x     y    
#>   <lgl> <lgl>
#> 1 NA    NA   
#> 
#> $result
#>   alpha   p1   p2  n1  n2  DF    pd         se critical        ME      lower
#> 1  0.01 0.38 0.51 100 200 298 -0.13 0.06004582 2.326348 0.1396875 -0.2696875
#>         upper                       CI   ppooled   sepooled         z
#> 1 0.009687456 -0.13 [99CI -0.27; 0.01] 0.4666667 0.06110101 -2.127624
#>       pvalue alternative
#> 1 0.01668412        less
#> 
#> $call
#> propCI(n1 = 100, n2 = 200, p1 = 0.38, p2 = 0.51, alpha = 0.01, 
#>     alternative = "less")
#> 
#> attr(,"measure")
#> [1] "propdiff"